--- src/Sulu/Bundle/AdminBundle/Resources/js/services/Initializer/index.js
@@ -1,4 +1,0 @@
-// @flow
-import initializer from './Initializer';
-
-export default initializer;
