+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/initializer/README.md
