+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/initializer/index.js
@@ -1,0 +1,4 @@
+// @flow
+import initializer from './initializer';
+
+export default initializer;
