+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/initializer/initializer.js
