+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/initializer/types.js
