+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/List/tests/registries/ListToolbarActionRegistry.test.js
@@ -2,7 +2,7 @@
 import listToolbarActionRegistry from '../../registries/listToolbarActionRegistry';
 import AbstractListToolbarAction from '../../toolbarActions/AbstractListToolbarAction';
 
-jest.mock('../../../../services/Initializer', () => jest.fn());
+jest.mock('../../../../services/initializer', () => jest.fn());
 jest.mock('../../toolbarActions/DeleteToolbarAction', () => jest.fn());
 
 beforeEach(() => {
